+++ app/src/main/java/ru/geekbrains/SettingsActivity.java
@@ -1,0 +1,62 @@
+package ru.geekbrains;
+
+import androidx.appcompat.app.AppCompatActivity;
+
+import android.content.Intent;
+import android.os.Bundle;
+import android.util.Log;
+import android.view.View;
+import android.widget.AdapterView;
+import android.widget.ArrayAdapter;
+import android.widget.Spinner;
+import android.widget.Toast;
+
+import java.util.ArrayList;
+
+public class SettingsActivity extends AppCompatActivity implements AdapterView.OnItemSelectedListener {
+    private final String TAG = "Activity";
+    Spinner spinner;
+    ArrayList spinnerArrayList;
+    ArrayAdapter spinnerAdapter;
+    String city;
+
+
+    @Override
+    protected void onCreate(Bundle savedInstanceState) {
+        super.onCreate(savedInstanceState);
+        setContentView(R.layout.activity_settings);
+
+
+        spinner = findViewById(R.id.spinner);
+        spinner.setOnItemSelectedListener(this);
+        spinnerArrayList = new ArrayList();
+
+        spinnerArrayList.add("Moscow");
+        spinnerArrayList.add("Tokyo");
+        spinnerArrayList.add("Paris");
+
+        spinnerAdapter = new ArrayAdapter(this, android.R.layout.simple_spinner_item, spinnerArrayList);
+        spinnerAdapter.setDropDownViewResource(android.R.layout.simple_spinner_item);
+        spinner.setAdapter(spinnerAdapter);
+    }
+
+
+
+    public void onClick(View view) {
+        Intent intent = new Intent(SettingsActivity.this, MainActivity.class);
+        intent.putExtra("city", city);
+        startActivity(intent);
+    }
+
+    @Override
+    public void onItemSelected(AdapterView<?> adapterView, View view, int i, long l) {
+        city = spinner.getSelectedItem().toString();
+    }
+
+    @Override
+    public void onNothingSelected(AdapterView<?> adapterView) {
+
+    }
+
+
+}
